requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Data & Analysis
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.35

# Technical Indicators
pandas-ta==0.3.14b0

# Statistics
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0

# Date/Time
pytz==2023.3
